Modify GMN to Canonical Form representations
Changed the logic of the AttributedGraphDataset to generate GMN pairs
based upon a briges associated Canonical Form representation
e.g (G1, CFG1). The AttributedGraphDataset loads two datasets,
a matching dataset which is the synthetic learning data and
a canonical dataset which contains canonical form representation of the
bridges in in the synthetic dataset.